api/requirements.txt
django>=4.2
djangorestframework
drf-yasg
mysqlclient
pandas
plotly
gunicorn
